+++ SpaceLaunches/Application/SceneDelegate.swift
@@ -10,13 +10,11 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         window = UIWindow(windowScene: windowScene)
         window?.makeKeyAndVisible()
-        window?.rootViewController = getLaunchListViewController()
-    }
 
-    private func getLaunchListViewController() -> LaunchListViewController {
-        let launchService = LaunchLibrary2Service()
-        let viewModel = LaunchListViewModel(service: launchService)
-        let viewController = LaunchListViewController(viewModel: viewModel)
-        return viewController
+        let navigationController = AppNavigationController()
+        let coordinator = LaunchListCoordinator(navigationController: navigationController)
+        coordinator.start()
+
+        window?.rootViewController = navigationController
     }
 }